Move AvailableStreams as separate container
